--- a/python/cusignal/_spectral.py
+++ b/python/cusignal/_spectral.py
@@ -425,7 +425,7 @@ def precompile_kernels(dtype=None, backend=None, k_type=None):
             _populate_kernel_cache(d, b, k)
 
 
-def _lombscargle(x, y, freqs, pgram, y_dot, cp_stream, use_numba):
+def _lombscargle(x, y, freqs, pgram, y_dot, cp_stream, autosync, use_numba):
 
     device_id = cp.cuda.Device()
     numSM = device_id.attributes["MultiProcessorCount"]
@@ -444,3 +444,6 @@ def _lombscargle(x, y, freqs, pgram, y_dot, cp_stream, use_numba):
     )
 
     kernel(x, y, freqs, pgram, y_dot)
+
+    if autosync is True:
+        cp_stream.synchronize()