--- a/python/cusignal/_upfirdn.py
+++ b/python/cusignal/_upfirdn.py
@@ -578,8 +578,9 @@ class _UpFIRDn(object):
         self,
         x,
         axis=-1,
-        cp_stream=cp.cuda.stream.Stream(null=True),
-        use_numba=False,
+        cp_stream,
+        autosync,
+        use_numba,
     ):
         """Apply the prepared filter to the specified axis of a nD signal x"""
 
@@ -637,5 +638,8 @@ class _UpFIRDn(object):
             padded_len,
             out,
         )
+
+        if autosync is True:
+            cp_stream.synchronize()
 
         return out