--- a/python/cusignal/spectral_analysis/__init__.py
+++ b/python/cusignal/spectral_analysis/__init__.py
@@ -19,5 +19,5 @@ from cusignal.spectral_analysis.spectral import (
     spectrogram,
     lombscargle,
     vectorstrength,
-    stft
+    stft,
 )